langgraph_tutorials/requirements.txt
# LangGraph 1.0.x 核心依赖
langgraph>=0.3.0
langchain>=1.0.0
langchain-core>=0.3.0
langchain-openai>=0.3.0

# 可视化工具
langgraph-cli>=0.1.0

# 工具
python-dotenv>=1.0.0
pydantic>=2.0.0
